+++ src/js/util.js
@@ -65,19 +65,33 @@ exports.sanitize = function (jsString) {
   function next()  { return jsString.charAt(i + 1); }
   function prev()  { return jsString.charAt(i - 1); }
 
+  function isWhiteSpace(c) {
+    return c === ' ' || c === '\n' || c === '\r' || c === '\t';
+  }
+
   // get the last parsed non-whitespace character
   function lastNonWhitespace () {
     var p = chars.length - 1;
 
     while (p >= 0) {
       var pp = chars[p];
-      if (pp !== ' ' && pp !== '\n' && pp !== '\r' && pp !== '\t') { // non whitespace
+      if (!isWhiteSpace(pp)) {
         return pp;
       }
       p--;
     }
 
     return '';
+  }
+
+  // get at the first next non-white space character
+  function nextNonWhiteSpace() {
+    var iNext = i + 1;
+    while (iNext < jsString.length && isWhiteSpace(jsString[iNext])) {
+      iNext++;
+    }
+
+    return jsString[iNext];
   }
 
   // skip a block comment '/* ... */'
@@ -166,7 +180,7 @@ exports.sanitize = function (jsString) {
     }
     else if (c === '\u00A0' || (c >= '\u2000' && c <= '\u200A') || c === '\u202F' || c === '\u205F' || c === '\u3000') {
       // special white spaces (like non breaking space)
-      chars.push(' ')
+      chars.push(' ');
       i++
     }
     else if (c === quote) {
@@ -183,6 +197,10 @@ exports.sanitize = function (jsString) {
     }
     else if (c === quoteDblLeft) {
       parseString(quoteDblRight);
+    }
+    else if (c === ',' && [']', '}'].indexOf(nextNonWhiteSpace()) !== -1) {
+      // skip trailing commas
+      i++;
     }
     else if (/[a-zA-Z_$]/.test(c) && ['{', ','].indexOf(lastNonWhitespace()) !== -1) {
       // an unquoted object key (like a in '{a:2}')